feat: optimizations based on profile runs in hashing and work coordinator. test updates and correctness fixes

+++ tests/unit/daemon/onnx_model_pool_catch2_test.cpp
@@ -295,7 +295,7 @@ TEST_CASE_METHOD(OnnxModelPoolFixture, "OnnxModelPool: LRU eviction", "[daemon]"
     // This test verifies the API exists and doesn't crash
     auto stats = pool_->getStats();
     CHECK(stats.loadedModels >= 0);
-    CHECK(stats.loadedModels <= static_cast<int>(config_.maxLoadedModels));
+    CHECK(stats.loadedModels <= config_.maxLoadedModels);
 }
 
 // Test memory limit enforcement
